King, Rook, Bishop, Queen
