+++ src/app/infrastructure/bishop.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { Cord, PossibleCords } from '../domain/basicChessTypes';
 
-export const getDirections = (cord: Cord): PossibleCords[] => {
+export const getBishopDirections = (cord: Cord): PossibleCords[] => {
     const { x, y } = cord;
     const directions = [
         ..._.zip(_.range(x + 1, 8, 1), _.range(y + 1, 8, 1)),
